+++ mail-server/clamav.nix
@@ -14,19 +14,17 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program. If not, see <http://www.gnu.org/licenses/>
 
-{ config, pkgs, lib, ... }:
+{ config, pkgs, lib, options, ... }:
 
 let
   cfg = config.mailserver;
 in
 {
   config = lib.mkIf (cfg.enable && cfg.virusScanning) {
-    services.clamav.daemon.enable = true;
+    services.clamav.daemon = {
+      enable = true;
+      settings.PhishingScanURLs = "no";
+    };
     services.clamav.updater.enable = true;
-
-    services.clamav.daemon.extraConfig = ''
-      PhishingScanURLs no
-    '';
   };
 }
-