+++ mail-server/opendkim.nix
@@ -29,7 +29,7 @@ let
       dkim_txt = "${cfg.dkimKeyDirectory}/${dom}.${cfg.dkimSelector}.txt";
     in
         ''
-          if [ ! -f "${dkim_key}" ] || [ ! -f "${dkim_txt}" ]
+          if [ ! -f "${dkim_key}" ]
           then
               ${pkgs.opendkim}/bin/opendkim-genkey -s "${cfg.dkimSelector}" \
                                                    -d "${dom}" \
@@ -37,15 +37,16 @@ let
                                                    --directory="${cfg.dkimKeyDirectory}"
               mv "${cfg.dkimKeyDirectory}/${cfg.dkimSelector}.private" "${dkim_key}"
               mv "${cfg.dkimKeyDirectory}/${cfg.dkimSelector}.txt" "${dkim_txt}"
+              chmod 644 "${dkim_txt}"
               echo "Generated key for domain ${dom} selector ${cfg.dkimSelector}"
           fi
         '';
   createAllCerts = lib.concatStringsSep "\n" (map createDomainDkimCert cfg.domains);
 
-  keyTable = pkgs.writeText "opendkim-KeyTable" 
-    (lib.concatStringsSep "\n" (lib.flip map cfg.domains 
+  keyTable = pkgs.writeText "opendkim-KeyTable"
+    (lib.concatStringsSep "\n" (lib.flip map cfg.domains
       (dom: "${dom} ${dom}:${cfg.dkimSelector}:${cfg.dkimKeyDirectory}/${dom}.${cfg.dkimSelector}.key")));
-  signingTable = pkgs.writeText "opendkim-SigningTable" 
+  signingTable = pkgs.writeText "opendkim-SigningTable"
     (lib.concatStringsSep "\n" (lib.flip map cfg.domains (dom: "${dom} ${dom}")));
 
   dkim = config.services.opendkim;
@@ -59,7 +60,7 @@ in
         keyPath = cfg.dkimKeyDirectory;
         domains = "csl:${builtins.concatStringsSep "," cfg.domains}";
         configFile = pkgs.writeText "opendkim.conf" (''
-          Canonicalization relaxed/simple
+          Canonicalization ${cfg.dkimHeaderCanonicalization}/${cfg.dkimBodyCanonicalization}
           UMask 0002
           Socket ${dkim.socket}
           KeyTable file:${keyTable}