+++ mail-server/users.nix
@@ -21,7 +21,7 @@ with config.mailserver;
 let
   vmail_user = {
     name = vmailUserName;
-    isNormalUser = false;
+    isSystemUser = true;
     uid = vmailUID;
     home = mailDirectory;
     createHome = true;
@@ -33,6 +33,9 @@ let
     #!${pkgs.stdenv.shell}
 
     set -euo pipefail
+
+    # Prevent world-readable paths, even temporarily.
+    umask 007
 
     # Create directory to store user sieve scripts if it doesn't exist
     if (! test -d "${sieveDirectory}"); then